requirements_personal.txt
bootstrap4==0.1.0
Django==2.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-password-reset==2.0
mysql-connector-python==8.0.13
Pillow==6.0.0
protobuf==3.6.1
pytz==2018.5
six==1.11.0
virtualenv==16.4.3
